% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter.hyperSpec}
\alias{filter.hyperSpec}
\title{filter hyperSpec object}
\usage{
\method{filter}{hyperSpec}(.data, ..., .preserve = FALSE)
}
\arguments{
\item{.data}{the hyperSpec object}

\item{...}{conditions, see \code{\link[dplyr:filter ]{dplyr::filter ()}}}

\item{.preserve}{recalculate grouping structure, see \code{\link[dplyr:filter ]{dplyr::filter ()}}}
}
\value{
hyperSpec object with rows (spectra) matching the conditions (logical predicates) in \code{...}
}
\description{
\code{filter}ing extra data columns works smoothly,
but filtering on the spectra matrix needs some attention:
the filtering expression must return a logical vector with one value per spectrum,
see examples.
}
\examples{
filter (flu, c < 0.2)
flu \%>\% filter (c < 0.2)

# filtering the spectra matrix:
# remove all spectra with NAs
tmp <- flu
tmp [[3:4,, 300 ~ 450]] <- NA
tmp [[6]] <- NA

tmp \%>\% filter (!any_wl (is.na (spc)))
tmp \%>\% filter (all_wl (!is.na (spc))) # the same

remove spectra that contain only NAs:
tmp \%>\% filter (!all_wl (is.na (spc)))
tmp \%>\% filter (any_wl (!is.na (spc))) # the same
}
