% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unittest.R
\name{hyperSpec.dplyr.unittest}
\alias{hyperSpec.dplyr.unittest}
\title{cbsvm unit tests}
\usage{
hyperSpec.dplyr.unittest()
}
\value{
Invisibly returns a data.frame with the test results
}
\description{
If \code{\link[testthat:testthat]{testthat::testthat()}} is available, run the unit tests and
display the results.
}
\examples{

hyperSpec.dplyr.unittest ()

}
\author{
Claudia Beleites
}
\keyword{programming}
\keyword{utilities}
